requirements.txt
yt-dlp
youtube-transcript-api
python-docx
